+++ nym-vpn-android/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2026.12.4] - 2026-09-16
+
 ## [2026.12.4] - 2026-08-27
 
 ### Added